+++ src/sources/mod.rs
@@ -69,6 +69,7 @@ pub(crate) mod string;
 pub(crate) mod umls;
 pub(crate) mod uniprot;
 pub(crate) mod vaers;
+pub(crate) mod who_ivd;
 pub(crate) mod who_pq;
 pub(crate) mod wikipathways;